Make Go 1.2 requirement more noticeable.
It was pointed out in #76 that if you arrived to the Update section
of the README without seeing the Installation section, the requirement for
Go 1.2 is easy to miss.  This commit builds the requirement in the
Installation section and adds it to the Updating section as well to
hopefully make it more noticable.